Split example classes into separate files for PSR-1 compliance
- Split UploadHandler.php into three separate files
- Create UploadResult.php and UploadException.php
- Remove require_once, rely on PSR-4 autoloading
- Add proper PHPDoc type hints (positive-int, list<string>)
- Improve code organization and maintainability
