+++ tests/Example/UploadException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Koriym\FileUpload\Example;
+
+use RuntimeException;
+
+final class UploadException extends RuntimeException
+{
+}
